# packages/api/openapi.json
{
...
  "info": {
    "description": "API documentation for Musetric",
    "title": "Musetric API",
    "version": "contract"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/preview/{previewId}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "previewId",
            "required": true,
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "content": { "application/json": { "schema": {} } },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": { "message": { "type": "string" } },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": { "message": { "type": "string" } },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/project/status/stream": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "Server-Sent Events stream (text/event-stream) with JSON encoded project status updates and heartbeat events.",
                  "type": "string"
                }
              }
            },
            "description": "Server-Sent Events stream (text/event-stream) with JSON encoded project status updates and heartbeat events."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": { "message": { "type": "string" } },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": { "message": { "type": "string" } },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        }
      }
    },
...
  }
}